resigned website

+++ lib/formatter.ts
@@ -1,9 +1,9 @@
 export function formatDateString(dateString: string): string {
-  const date = new Date(dateString);
-  const options: Intl.DateTimeFormatOptions = {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  };
-  return date.toLocaleDateString("en-US", options);
+    const date = new Date(dateString);
+    const options: Intl.DateTimeFormatOptions = {
+        year: "numeric",
+        month: "short",
+        day: "numeric",
+    };
+    return date.toLocaleDateString("en-US", options);
 }
